requirements.txt
# Core dependencies
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Data processing
thefuzz>=0.20.0
python-Levenshtein>=0.21.0

# Optional: For better logging
rich>=13.0.0